Keep raw X in its own dtype; widen derived quantities back to float64
The blanket float32 conversion in acea8c1 hardcoded intermediate arrays
(mttkrp, Q_i, Y_i, W, factor inits, etc.) to float32 regardless of X's
actual dtype, losing precision even when X is float64. It also breaks
under nvmath's sparse matmul (merged from main), which requires all
matmul operands to share one exact dtype -- so a float64 X now fails
outright instead of silently upcasting.

Narrow the fix to only what's needed for GPU/nvmath compatibility: cast
just the dense operand of each X matmul/rmatmul to X's own dtype right
before the call (Omega, Q.T, Q_z in randomized_svd_right; C_op in
calc_W; H_T in parafac_update). Every other derived quantity reverts to
float64, restoring the previous numerical precision for float64
datasets while keeping float32 X unpromoted.

+++ parafac2/parafac2.py
@@ -152,8 +152,8 @@ def parafac2_init(
     )
 
     factors = [
-        np.ones((n_cond, rank), dtype=np.float32),
-        np.eye(rank, dtype=np.float32),
+        np.ones((n_cond, rank), dtype=np.float64),
+        np.eye(rank, dtype=np.float64),
         C,
     ]
     return factors, norm_tensor
